Fix histmatch for RealLike arrays
While formerly we could call the non-exported _histmatch, we were
missing an exported method for non-Colorant images. In implementing
this it was revealed that integer-valued arrays need special handling
so that lack of rounding doesn't cause InexactErrors.

This also introduces a couple of minor behavior changes:
- in constructing the lookup table, this switches to choosing the last
  (rather than the first) bin that has the target cumulative probability.
  This choice does a better job of ensuring that the returned pixel
  values match those present in the reference image (formerly we
  returned values corresponding to gaps in the histogram).
- this no longer zeros the first and last bin counts. I'm not sure why
  this was done previously, but in some cases it led to unnecessary
  scaling of the cumulative density because it effectively modified the
  number of pixels in the image.

Note a couple of tests fail, but the behavior they enforce may
not be desirable. I'll delete those in the next commit.